+++ tests/components/component.py
@@ -1,0 +1,3 @@
+class Component(object):
+    def __init__(self, driver):
+        self.driver = driver
