--- tests/first_test.py
@@ -1,31 +1,0 @@
-import os
-
-import unittest
-import urlparse
-
-from selenium.webdriver import DesiredCapabilities, Remote
-from selenium.webdriver.support.ui import WebDriverWait
-
-from mail_auth import Authentification
-
-class ExampleTest(unittest.TestCase):
-
-    def setUp(self):
-        browser = os.environ.get('BROWSER', 'CHROME')
-
-        self.driver = Remote(
-            command_executor='http://127.0.0.1:4444/wd/hub',
-            desired_capabilities=getattr(DesiredCapabilities, browser).copy()
-        )
-
-    def tearDown(self):
-        self.driver.quit()
-
-    def test(self):
-        prepare = Authentification(self.driver)
-        prepare.auth()
-
-        #should be replaced by wait
-        import time
-        time.sleep(5)
-        #end(should be replaced by wait)
