--- tests/mail_auth.py
@@ -1,58 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import os
-import urlparse
-
-class Page(object):
-    BASE_URL = 'https://mail.ru/'
-    PATH = ''
-
-    def __init__(self, driver):
-        self.driver = driver
-
-    def open(self):
-        url = urlparse.urljoin(self.BASE_URL, self.PATH)
-        self.driver.get(url)
-        self.driver.maximize_window()
-
-
-class AuthPage(Page):
-    PATH = ''
-
-    @property
-    def form(self):
-        return AuthForm(self.driver)
-
-class Component(object):
-    def __init__(self, driver):
-        self.driver = driver
-
-class AuthForm(Component):
-    LOGIN = '//input[@id="mailbox:login"]'
-    PASSWORD = '//input[@id="mailbox:password"]'
-    SUBMIT = '//input[@class="o-control"]'
-
-    def set_login(self, login):
-        self.driver.find_element_by_xpath(self.LOGIN).send_keys(login)
-
-    def set_password(self, pwd):
-        self.driver.find_element_by_xpath(self.PASSWORD).send_keys(pwd)
-
-    def submit(self):
-        self.driver.find_element_by_xpath(self.SUBMIT).click()
-
-class Authentification(object):
-    USEREMAIL = 'ttexnopark@mail.ru'
-    PASSWORD = os.environ['PASSWORD']
-
-    def __init__(self, driver):
-        self.driver = driver
-
-    def auth(self):
-        auth_page = AuthPage(self.driver)
-        auth_page.open()
-
-        auth_form = auth_page.form
-        auth_form.set_login(self.USEREMAIL)
-        auth_form.set_password(self.PASSWORD)
-        auth_form.submit()
