+++ tests/pages/main_page.py
@@ -1,0 +1,10 @@
+import sys
+
+from page import Page
+from tests.components.sidebar import Sidebar
+
+class MainPage(Page):
+    @property
+    def sidebar(self):
+        return Sidebar(self.driver)
+
